Sort out debug daily reminder option to run with AlarmManager
