+++ android_mobile/src/main/AndroidManifest.xml
@@ -182,6 +182,8 @@
             android:name="com.alexstyl.specialdates.wear.WearSyncService"
             android:exported="false" />
 
+        <receiver android:name=".dailyreminder.DailyReminderReceiver" />
+        
         <receiver android:name="com.alexstyl.specialdates.DeviceConfigurationUpdatedReceiver">
             <intent-filter>
                 <action android:name="android.intent.action.LOCALE_CHANGED" />
@@ -242,6 +244,5 @@
             android:name="com.google.firebase.messaging.default_notification_channel_id"
             android:value="default_channel_id" />
 
-        <receiver android:name=".dailyreminder.DailyReminderReceiver" />
     </application>
 </manifest>
